GET students via routes
