+++ students/README.md
@@ -122,7 +122,7 @@ dropdb lambda
 Add the following endpoints.
 
 - `[POST] /students` This route should save a new student to the database.
-- `[GET] /students` This route will return an array of all students.
+- [x] `[GET] /students` This route will return an array of all students.
 - `[GET] /students/:id` This route will return the student with the matching `id`.
 - `[PUT] /students/:id` This route will update the student with the matching `id` using information sent in the body of the request.
 - `[DELETE] /students/:id` This route should delete the specified student.
