+++ students/db/seeds/01_cohorts.js
@@ -1,13 +1,17 @@
-const faker = require('faker')
-
-const createFakeCohort = () => ({
-  name: faker.commerce.department()
-})
-
 exports.seed = async function(knex, Promise) {
-  const fakeCohorts = []
-  for (let i = 0; i < 10; i++) {
-    fakeCohorts.push(createFakeCohort())
-  }
-  return await knex('cohorts').insert(fakeCohorts)
+  const cohorts = [
+    {
+      cohortId: 1,
+      name: 'Web'
+    },
+    {
+      cohortId: 2,
+      name: 'iOS'
+    },
+    {
+      cohortId: 3,
+      name: 'UX'
+    }
+  ]
+  return await knex('cohorts').insert(cohorts)
 }
