+++ students/index.js
@@ -1,20 +1,13 @@
 const express = require('express')
-const knex = require('./db/knex')
+const cohortRoutes = require('./routes/cohortRoutes')
+const studentRoutes = require('./routes/studentRoutes')
 
 const app = express() // initialize Express
 const port = process.env.PORT || 1000
 app.listen(port, () => console.log(`Server is running 🏃`))
 
-// route handlers
-const getCohorts = async (req, res) => {
-  try {
-    const cohorts = await knex.select().from('cohorts')
-    return res.status(200).json(cohorts)
-  } catch (error) {
-    console.error(error)
-    res.status(500).json({ error, msg: 'Error getting cohorts' })
-  }
-}
+app.use(express.json())  // parser for incoming data
 
 // routes
-app.get('/', getCohorts)
+app.use('/', cohortRoutes)
+app.use('/students', studentRoutes)
